@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <RootNamespace>CalcmenuAPI</RootNamespace>
     <AssemblyName>CalcmenuAPI.Models</AssemblyName>
     <OptionStrict>Off</OptionStrict>
