@@ -1,5 +1,5 @@
 using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using log4net;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
