@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Data;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using log4net;
 using Microsoft.AspNetCore.Mvc;
 using static EgsData.modGlobalDeclarations; // ConnectionString
